+++ src/services/gprc/world_engine_client_grpc.service.ts
@@ -5,6 +5,7 @@ import {
   CapabilitiesRequest,
   ClientRequest,
   LoadSceneRequest,
+  LoadSceneResponse,
   UserRequest,
 } from '@proto/world-engine_pb';
 import { promisify } from 'util';
@@ -13,6 +14,7 @@ import { Config } from '../../common/config';
 import { CLIENT_ID } from '../../common/constants';
 import { Awaitable } from '../../common/data_structures';
 import { grpcOptions } from '../../common/helpers';
+import { Logger } from '../../common/logger';
 import { SessionToken } from '../../entities/session_token.entity';
 
 export interface LoadSceneProps {
@@ -32,13 +34,19 @@ export interface SessionProps {
 
 export class WorldEngineClientGrpcService {
   private readonly config = Config.getInstance();
+  private readonly address = this.config.getEngineHost();
+  private readonly ssl = this.config.getEngineSsl();
+  private readonly credentials = this.ssl
+    ? credentials.createSsl()
+    : credentials.createInsecure();
+  private readonly grpcOptions = { ...grpcOptions };
   private readonly client = new WorldEngineClient(
-    this.config.getEngineHost(),
-    this.config.getEngineSsl()
-      ? credentials.createSsl()
-      : credentials.createInsecure(),
-    { ...grpcOptions },
+    this.address,
+    this.credentials,
+    this.grpcOptions,
   );
+
+  private logger = Logger.getInstance();
 
   public async loadScene(props: LoadSceneProps) {
     const { name, sessionToken, user, capabilities } = props;
@@ -55,20 +63,47 @@ export class WorldEngineClientGrpcService {
       new ClientRequest().setId(props.client?.getId() || CLIENT_ID),
     );
 
+    const metadata = this.getMetadata(sessionToken);
     const finalRequest = props.setLoadSceneProps
       ? props.setLoadSceneProps(request)
       : request;
 
-    return promisify(this.client.loadScene.bind(this.client))(
-      finalRequest,
-      this.getMetadata(sessionToken),
-    );
+    const response: LoadSceneResponse = await promisify(
+      this.client.loadScene.bind(this.client),
+    )(finalRequest, metadata);
+
+    this.logger.debug({
+      action: 'Load scene',
+      data: {
+        address: this.address,
+        ssl: this.ssl,
+        grpcOptions: this.grpcOptions,
+        metadata: metadata.toJSON(),
+        request: request.toObject(),
+        response: response.toObject(),
+      },
+      sessionId: sessionToken.sessionId,
+    });
+
+    return response;
   }
 
   public session(props: SessionProps) {
     const { sessionToken, onDisconnect, onError, onMessage } = props;
 
-    const connection = this.client.session(this.getMetadata(sessionToken));
+    const metadata = this.getMetadata(sessionToken);
+    const connection = this.client.session(metadata);
+
+    this.logger.debug({
+      action: 'Open session',
+      data: {
+        address: this.address,
+        ssl: this.ssl,
+        grpcOptions: this.grpcOptions,
+        metadata: metadata.toJSON(),
+      },
+      sessionId: sessionToken.sessionId,
+    });
 
     if (onMessage) {
       connection.on('data', onMessage);